requirements.txt
fastapi
uvicorn
pydantic
langgraph
crewai
litellm
crewai[tools]
langchain-groq
requests
python-dotenv
sqlalchemy 
psycopg2-binary